[PSym] refactor: separate out symbolic and explicit
