+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/utils/monitor/TimedCall.java
@@ -3,7 +3,6 @@ package psym.utils.monitor;
 import psym.runtime.scheduler.SearchScheduler;
 import psym.utils.exception.BugFoundException;
 import psym.utils.exception.MemoutException;
-import psym.utils.monitor.MemoryMonitor;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.TimeoutException;
